% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cndcR.R
\docType{data}
\name{Data}
\alias{Data}
\title{Case study on maize CNDC dataset.}
\format{
A data frame with 1150 rows and 39 variables
}
\usage{
data(Data)
}
\description{
The data were extracted from the global synthesis database developed by Ciampitti et al. (2022b) for a global NNI
initiative across species. From the cited database, we utilized a subset for maize crop species, including 1150 pairs
of observations for W and N% published in eleven studies. The variables of interest for the present study are:
}
\details{
\itemize{
  \item Id. Number of publication.
  \item Paper. Publication of reference.
  \item Species. Crop species name.
  \item Year. Year of publication.
  \item Country.
  \item Sites. Location of the study.
  \item e. Experimental design of the study.
  \item Repetitions. Number of replicates
  \item Nrates. N fertilization supply treatment in kg ha-1
  \item W_kg_ha. Shoot biomass W in kg ha-1
  \item Nupt_kg_ha. Shoot N uptake in kg ha-1
  \item Na. Shoot N concentration in %
  \item Samp. Sampling time
  \item Site_year. Id for the GxExM combination
  \item Na_SE. standard deviation for shoot N%
  \item W_SE. standard deviation for shoot W biomass
}
}
\keyword{datasets}
