% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Freq_method.R
\name{fig1a}
\alias{fig1a}
\title{Plot figure 1a from Fernandez et al. (2022)}
\usage{
fig1a()
}
\description{
This function returns figure 1a in Fernandez et al.(2022): Frequency distribution of statistical
method used for fitting CNDC in literature.
}
