% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Freq_Ndeterm.R
\name{fig1b}
\alias{fig1b}
\title{Plot figure 1b from Fernandez et al. (2022)}
\usage{
fig1b()
}
\description{
This function returns figure 1b in Fernandez et al.(2022): Frequency distribution of N determination method used for fitting CNDC
in literature.
}
