% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BoxPlot_a.R
\name{fig1c}
\alias{fig1c}
\title{Plot figure 1c from Fernandez et al. (2022)}
\usage{
fig1c()
}
\description{
This function returns figure 1c in Fernandez et al.(2022): Boxplots and density plots for the
distribution of the a (A1) parameter of the CNDC for C3 (red color) and C4 (blue color) species
in studies from literature.
}
