% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BoxPlot_b.R
\name{fig1d}
\alias{fig1d}
\title{Plot figure 1d from Fernandez et al. (2022)}
\usage{
fig1d()
}
\description{
This function returns figure 1d in Fernandez et al.(2022): Boxplots and density plots for the distribution
of the b (A2) parameter of the CNDC for C3 (red color) and C4 (blue color) species in studies from literature.
}
