% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Freq_exps.R
\name{fig2a}
\alias{fig2a}
\title{Plot figure 2a from Fernandez et al. (2022)}
\usage{
fig2a()
}
\description{
This function returns figure 2a in Fernandez et al.(2022): Frequency distribution of number of
experiments used for fitting CNDC in literature.
}
