% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sens_analysis_1.R
\name{fig2b_c}
\alias{fig2b_c}
\title{Plot figure 2b and 2c from Fernandez et al. (2022)}
\usage{
fig2b_c()
}
\description{
This function returns figure 2b and 2c in Fernandez et al.(2022): Sensitivity analyses for the number of
experiments used to estimate a and b parameters of the CNDC in a case-study for maize, presenting bias
and uncertainty (precision).
}
