% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Freq_rates.R
\name{fig3a}
\alias{fig3a}
\title{Plot figure 3a from Fernandez et al. (2022)}
\usage{
fig3a()
}
\description{
This function returns figure 3a in Fernandez et al.(2022): Frequency distribution of number of N-rates used for fitting CNDC
in literature.
}
