% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Freq_samps.R
\name{fig3b}
\alias{fig3b}
\title{Plot figure 3b from Fernandez et al. (2022)}
\usage{
fig3b()
}
\description{
This function returns figure 3b in Fernandez et al.(2022): Frequency distribution of number of
samplings used for fitting CNDC in literature.
}
