% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sensAnalysis_2.R
\name{fig3c_d}
\alias{fig3c_d}
\title{Plot figure 3c and 3d from Fernandez et al. (2022)}
\usage{
fig3c_d()
}
\description{
This function returns figure 3c and 3d in Fernandez et al.(2022): Sensitivity analyses for the number of
N rates used to estimate a and b parameters of the CNDC in a case-study for maize, presenting bias
and uncertainty (precision).
}
