% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sensAnalysis_3.R
\name{fig3e_f}
\alias{fig3e_f}
\title{Plot figure 3e and 3f from Fernandez et al. (2022)}
\usage{
fig3e_f()
}
\description{
This function returns figure 3e and 3f in Fernandez et al.(2022): Sensitivity analyses for the
number of sampling times used to estimate a and b parameters of the CNDC in a case-study for
maize, presenting bias and uncertainty (precision).
}
