% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Wrange.R
\name{fig4a}
\alias{fig4a}
\title{Plot figure 4a from Fernandez et al. (2022)}
\usage{
fig4a()
}
\description{
This function returns figure 4a in Fernandez et al.(2022): Regression: Range of biomass of the CNDC across crop species.
}
