% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Freq_lowWcriteria.R
\name{fig4b}
\alias{fig4b}
\title{Plot figure 4b from Fernandez et al. (2022)}
\usage{
fig4b()
}
\description{
This function returns figure 4b in Fernandez et al.(2022): Frequency distribution of reported low W criteria for fitting CNDC
in literature.
}
