% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Quantile_Reg.R
\name{fig4c}
\alias{fig4c}
\title{Plot figure 4c from Fernandez et al. (2022)}
\description{
This function returns figure 4c in Fernandez et al.(2022): Regression: minW vs. b coefficient of the CNDC.
}
