% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantileReg_b.R
\name{fig4d}
\alias{fig4d}
\title{Plot figure 4d from Fernandez et al. (2022)}
\usage{
fig4d()
}
\value{
figQuantile_Reg: ggplot object figure
}
\description{
This function returns figure 4d in Fernandez et al.(2022): Regression: minW vs. b coefficient of
the CNDC.
}
