% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sensAnalysis_4and5.R
\name{fig5b_c}
\alias{fig5b_c}
\title{Plot figure 5b and 5c from Fernandez et al. (2022)}
\usage{
fig5b_c()
}
\description{
This function returns figure 5b and 5c in Fernandez et al.(2022): SHistograms and density plots of the
posterior distributions for a (A1) and b (A2) parameters of the critical nitrogen dilution curve (CNDC)
of maize using four types of datasets: (i) All data or (ii) only Wmax plateau achieved data, and
(iii) unweighted or (iv) variance-weighted data.
}
