% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_sensAnalysis_4and5.R
\name{table1}
\alias{table1}
\title{Retrieve table 1 from Fernandez et al. (2022)}
\usage{
table1()
}
\description{
This function returns values for table 1 in Fernandez et al.(2022): Evaluation of four critical nitrogen dilution curves (CNDC)
models fitted with the Bayesian approach using four types of datasets: (All data or only Wmax achieved data) ✕
(unweighted or variance-weighted data model).
}
\details{
The four models were compared based on model parameters:
1.  Posterior distributions for a (A1) and b (A2) are averaged to obtain posterior medians and credibility intervals from all the samples.
2.  Then, they are used to estimate NNI values across an independent validation set of observations and compared against the reference
curve by Plenet and Lemaire (2000).
3.  Error metrics for the agreement between NNI estimated and reference are calculated using the metrica:: library.
4.  Deviance information criterion (DIC) is retrieved for each model.
}
\examples{
\dontrun{
Compare DIC of models ---------------------------------------------
dicSens_4and5 <- data_frame(Method = c("Unweighted-Wmax", "Unweighted-Wall", "Weighted-Wmax", "Weighted-Wall"))  \%>\%
  mutate(DIC = case_when(
    Method == "Unweighted-Wmax" ~ dic_4[[1]][[1]]  \%>\%  sum(),
    Method == "Unweighted-Wall" ~ dic_4[[2]][[1]]  \%>\%  sum(),
    Method == "Weighted-Wmax" ~ dic_5[[1]][[1]]  \%>\%  sum(),
    Method == "Weighted-Wall" ~ dic_5[[2]][[1]]  \%>\%  sum(),
    TRUE ~ NA_real_
  ))  \%>\%
  mutate_if(is.double, .funs = ~ round(digits = 0, x = .))
}
}
