% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cndcR.R
\docType{data}
\name{validationSet}
\alias{validationSet}
\title{Validation data of four studies in maize used in Table 1 comparing CNDC estimations}
\format{
A data frame
}
\usage{
data(validationSet)
}
\description{
Validation data of four studies in maize used in Table 1 comparing CNDC estimations
}
\keyword{datasets}
